move getIconSize to utils
