+++ packages/react-components/src/components/Icons/SvgCheckCircleIcon/SvgCheckCircleIcon.tsx
@@ -1,7 +1,7 @@
 /* This component implements the Circle Check icon from Font Awesome: https://fontawesome.com/icons/circle-check */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgCheckCircleIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
