+++ packages/react-components/src/components/Icons/SvgCheckIcon/SvgCheckIcon.tsx
@@ -1,7 +1,7 @@
 /* This component implements the Check icon from Font Awesome: https://fontawesome.com/icons/check */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgCheckIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
