+++ packages/react-components/src/components/Icons/SvgChevronDownIcon/SvgChevronDownIcon.tsx
@@ -1,7 +1,7 @@
 /* This component implements the Chevron Down icon from Font Awesome: https://fontawesome.com/icons/chevron-down */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgChevronDownIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
