+++ packages/react-components/src/components/Icons/SvgChevronRightIcon/SvgChevronRightIcon.tsx
@@ -1,7 +1,7 @@
 /* The component implements the Chevron Right icon from Font Awesome: https://fontawesome.com/icons/chevron-right */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgChevronRightIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
