+++ packages/react-components/src/components/Icons/SvgChevronUpIcon/SvgChevronUpIcon.tsx
@@ -1,7 +1,7 @@
 /* This component implements the Chevron Up icon from Font Awesome: https://fontawesome.com/icons/chevron-up */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgChevronUpIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
