+++ packages/react-components/src/components/Icons/SvgExclamationIcon/SvgExclamationIcon.tsx
@@ -1,7 +1,7 @@
 /* This component implements the Triangle Exclamation icon from Font Awesome: https://fontawesome.com/icons/triangle-exclamation */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgExclamationIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
