+++ packages/react-components/src/components/Icons/SvgMinusIcon/SvgMinusIcon.tsx
@@ -1,7 +1,7 @@
 /* This component implements the Minus icon from Font Awesome: https://fontawesome.com/icons/minus */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgMinusIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
