+++ packages/react-components/src/components/Icons/SvgReactAriaIcon/SvgReactAriaIcon.tsx
@@ -1,7 +1,7 @@
 /* React Aria logo icon, used in Storybook docs. Not intended for general use */
 
 import { useId } from "react-aria/useId";
-import getIconSize, { IconSize } from "../../../hooks/getIconSize";
+import getIconSize, { IconSize } from "../../../utils/getIconSize";
 
 export interface SvgReactAriaIconProps extends React.SVGProps<SVGSVGElement> {
   size?: IconSize;
